pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securescan"
version = "0.1.0"
description = "Automated security audit pipeline powered by LLM reasoning"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "gitpython>=3.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "tiktoken>=0.7",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
securescan = "securescan.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["securescan*"]
